refactor(frontend): isolate composer selection range mapping
